// frontend-management-pos/src/app/page/model/request/book-updated-request.model.ts
export interface BookUpdatedRequest {
    name?: string;
    publicId?: string|null;
    imgUrl?: string|null;
    quantity?: number;
    price?: number;
    publisher?: string|null;
    translator?: string|null;
    numOfPages?: number|null;
    publishedYear?: number|null;
    description?: string|null;
    authorId?: number;
    categoryId?: number;
}